bitcoin: refactor number encoding to Stack methods
Creates a new class Stack to house encodeNum and decodeNum for
converting these values into signed-magnitude representation.
